requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
textblob==0.18.0
pandas==2.1.4
nltk==3.8.1
google-generativeai==0.3.2
